change hover for icons, rename files temp
