+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import Header from './layouts/Header/Header';
-import Home from './pages/Home/Home';
+import Home from './pages/Home/Home2';
 import Skills from './pages/Skills/Skills';
 import Projects from './pages/Projects/Projects';
 import Footer from './layouts/Footer/Footer';
