+++ src/layouts/Header/Header.css
@@ -184,7 +184,7 @@
 }
 
 .contactIcon.sun:hover {
-    fill: #DAA520;
+    fill: #ac7e0c;
 }
 
 .contactIcon.moon:hover {
